Reorganize modules
* Move awsbatch and slurm specific model classes into new modules.
* Move schema classes in common between imagebuilder_schema and cluster_schema
  into a new common_schema module.
